feat: implement room management and player connection handling
